backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
numpy>=2.0.0    
scikit-learn==1.6.1
torch==2.3.0+cpu
torchvision==0.18.0+cpu
joblib==1.4.2
geopy==2.4.1
httpx==0.27.0
pandas==2.2.2
pyarrow==16.1.0
reportlab==4.2.0
pluscodes==1.0.0
